requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sse-starlette==2.1.3
pydantic==2.10.2
pydantic-settings==2.6.1
numpy==2.1.3
scipy==1.14.1
soundfile==0.12.1
ffmpeg-python==0.2.0
openai-whisper==20240930
whisperx==3.1.1
pyannote.audio==3.3.2
speechbrain==1.0.2
tqdm==4.67.1
orjson==3.10.12
python-dotenv==1.0.1
httpx==0.27.2
